module3-task1: * move common helpers/enums
