+++ src/common/enums/app.js
