--- src/common/enums/app/app.js
@@ -1,1 +1,0 @@
-export * from './app-route.enum';
