--- src/common/enums/offer/offers.js
@@ -1,2 +1,0 @@
-export * from './offer-city.enum';
-export * from './offer-type.enum';
