+++ src/common/enums/offers.js
@@ -7,4 +7,11 @@ const OfferCity = {
   DUSSELDORF: `Dusseldorf`,
 };
 
-export {OfferCity};
+const OfferType = {
+  APARTMENT: `apartment`,
+  ROOM: `room`,
+  HOUSE: `house`,
+  HOTEL: `hotel`,
+};
+
+export {OfferCity, OfferType};
