--- src/helpers/array/array.js
@@ -1,1 +1,0 @@
-export * from './get-random-item.helper';
