--- src/helpers/mocks/generate-offers.helper.js
@@ -1,9 +1,0 @@
-import {generateOffer} from '~/helpers/mocks/generate-offer.helper';
-
-const generateOffers = (count) => {
-  const offers = Array.from(new Array(count), generateOffer);
-
-  return offers;
-};
-
-export {generateOffers};
