--- src/helpers/mocks/mocks.js
@@ -1,3 +1,0 @@
-export * from './generate-offer.helper';
-export * from './generate-offers.helper';
-export * from './generate-random-id.helper';
