--- src/helpers/number/number.js
@@ -1,1 +1,0 @@
-export * from './get-random-number.helper';
