+++ src/helpers/string.js
